Extract presentation logic in SettingsPage to a custom hook
